Don't use proxy/almost empty index files
